requirements.txt
appnope==0.1.2
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
idna==3.3
ipykernel==6.6.0
ipython==7.30.1
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.21.4
pandas==1.3.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.23
ptyprocess==0.7.0
Pygments==2.10.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyzmq==22.3.0
requests==2.26.0
six==1.16.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
